pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tabify2"
version = "0.1.0"
description = "Tabify application"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "jinja2==3.1.2",
    "aiofiles==23.2.1",
    "email-validator==2.1.0.post1",
    "pydantic[email]==2.5.2",
    "python-slugify==8.0.1",
    "Pillow==10.1.0",
    "starlette==0.27.0",
    "itsdangerous==2.1.2",
    "requests==2.31.0",
    "alembic==1.12.1",
    "pytz==2023.3.post1",
    "qrcode==7.4.2",
    "httpx==0.27.0",
    "mido==1.2.10",
    "pretty-midi==0.2.10",
    "music21==9.1.0",
    "pyguitarpro==0.6",
    "uuid==1.30",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["*"]
